Backtracking works now.
Fixes #46
