+++ README.md
@@ -8,20 +8,20 @@
     7. Click on play
 
 # Implementing a Searcher
-    1. Create a class "yourSearcher" in
+    1. Create a class "YourSearcher" in
     src/main/java/com/treasure/hunt/strategy/searcher/impl/
     2. Implement to the interface "Searcher"
 
 # Implementing a Hider
-    1. Create a class "yourHider" in
+    1. Create a class "YourHider" in
     src/main/java/com/treasure/hunt/strategy/hider/impl/
     2. Implement to the interface "Hider"
     
 # Implementing a View
     Starting the GameManager with your View Class applied will
     run it **concurrent**!
-    1. Create a class "yourView" in
+    1. Create a class "YourView" in
     src/main/java/com/treasure/hunt/view/in_game/impl/
     2. Implement to the interface "View"
     3. Access the Products created after each step via
-    GameHistory.giveProductsCopy()
+    GameHistory.getGeometryItems()
