--- src/main/java/com/treasure/hunt/game/GameEvent.java
@@ -1,7 +1,0 @@
-package com.treasure.hunt.game;
-
-import com.treasure.hunt.strategy.hint.Hint;
-
-public class GameEvent<H extends Hint> {
-
-}
