+++ src/main/java/com/treasure/hunt/game/GameManager.java
@@ -1,67 +1,84 @@
 package com.treasure.hunt.game;
 
 import com.treasure.hunt.strategy.geom.GeometryItem;
-import com.treasure.hunt.strategy.geom.GeometryType;
 import com.treasure.hunt.strategy.hider.Hider;
-import com.treasure.hunt.strategy.hint.Hint;
-import com.treasure.hunt.strategy.searcher.Movement;
 import com.treasure.hunt.strategy.searcher.Searcher;
-import com.treasure.hunt.utils.JTSUtils;
-import com.treasure.hunt.utils.Requires;
 import com.treasure.hunt.view.in_game.View;
-import lombok.Getter;
-import org.locationtech.jts.geom.Coordinate;
-import org.locationtech.jts.geom.LineSegment;
-import org.locationtech.jts.geom.Point;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
-/**
- * This is the GameManager which should be started,
- * to start a normal game.
- */
-@Requires(hider = Hider.class, searcher = Searcher.class)
 public class GameManager {
 
-    protected List<View> view;
-    protected final GameHistory gameHistory = new GameHistory();
-
-    protected final Searcher searcher;
-    protected final Hider hider;
     /**
-     * Safe, whether the game is done or not.
+     * The {@link View} objects (being {@link Runnable})
      */
-    @Getter
-    protected boolean finished = false;
-    protected Hint lastHint;
-    protected Movement lastMovement;
-    protected Point searcherPos;
-    protected Point treasurePos;
+    private List<Runnable> views = new ArrayList<>();
+    /**
+     * Runs the {@link View} objects concurrently.
+     */
+    private ExecutorService executorService = Executors.newFixedThreadPool(8);
 
-    protected boolean firstMove = true;
-    protected int stepSim = 0;
-    protected int stepView = 0;
+    /**
+     * Locks GameHistory.products, such that it can only be accessed by one thread
+     * at the same time.
+     */
+    private List<Move> moves = Collections.synchronizedList(new ArrayList<>());
+    private GameEngine gameEngine;
 
-    public GameManager(Searcher searcher, Hider hider, List<View> view) {
-        this.searcher = searcher;
-        this.hider = hider;
-        this.view = view;
+    private int stepSim = 0;
+    private int stepView = 0;
+
+    /**
+     * @param searcherClass   (Sub-)class of {@link Searcher}
+     * @param hiderClass      (Sub-)class of {@link Hider}
+     * @param gameEngineClass (Sub-)class of {@link GameEngine}
+     * @param views           A list of {@link View} objects ({@link Runnable})
+     * @throws NoSuchMethodException     from {@link Class#getDeclaredConstructor(Class[])}
+     * @throws IllegalAccessException    from {@link java.lang.reflect.Constructor#newInstance(Object...)}
+     * @throws InvocationTargetException from {@link java.lang.reflect.Constructor#newInstance(Object...)}
+     * @throws InstantiationException    from {@link java.lang.reflect.Constructor#newInstance(Object...)}
+     */
+    public GameManager(Class<? extends Searcher> searcherClass, Class<? extends Hider> hiderClass, Class<? extends GameEngine> gameEngineClass, List<View> views)
+            throws NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException {
+
+        Searcher newSearcher = searcherClass.getDeclaredConstructor().newInstance();
+        Hider newHider = hiderClass.getDeclaredConstructor().newInstance();
+        GameEngine gameEngineInstance = gameEngineClass
+                .getDeclaredConstructor(Searcher.class, Hider.class)
+                .newInstance(newSearcher, newHider);
+        this.gameEngine = gameEngineInstance;
+
+        for (View view : views) {
+            registerListener(view);
+            view.init(this);
+        }
+
+        moves.add(gameEngine.init());
+        stepView++;
+        stepSim++;
+        runListeners();
     }
 
     /**
      * Works only for stepSim <= stepView
      */
     public void next() {
-        if (stepSim <= stepView) {
-            stepView++;
+        if (stepView <= stepSim) {
 
-            if (stepSim == stepView) {
-                move();
-            } else {
-                gameHistory.next();
+            if (stepView == stepSim) {
+                moves.add(gameEngine.move());
+                runListeners();
+                stepSim++;
             }
+            stepView++;
         }
+        runListeners();
     }
 
     /**
@@ -70,153 +87,30 @@ public class GameManager {
     public void previous() {
         if (stepView > 0) {
             stepView--;
-            gameHistory.previous();
         }
+        runListeners();
     }
 
     /**
-     * This simulates just one step of the simulation.
-     * The searcher begins since we want not force him,
-     * to take a initial hint, he eventually do not need,
-     * if he works randomized!
-     * <p>
-     * Updates the searchers position.
-     * <p>
-     * The first step of the searcher goes without an hint,
-     * the next will be with.
+     * @param runnable the views, which gets registered.
      */
-    public void move() {
-        if (finished) {
-            throw new IllegalStateException("Game is already finished");
-        }
-
-        // Searcher moves
-        if (firstMove) {
-            firstMove = false;
-            lastMovement = searcher.move();
-        } else {
-            lastMovement = searcher.move(lastHint);
-        }
-        assert (lastMovement != null);
-        assert (lastMovement.getPoints().size() != 0);
-        searcherPos = lastMovement.getEndPoint();
-
-        if (located(lastMovement.getPoints())) {
-            finish();
-            return;
-        } else {
-            lastHint = hider.move(lastMovement);
-        }
-        assert (lastHint != null);
-
-        if (!checkConsistency()) {
-            throw new IllegalStateException("Game is no longer consistent!");
-        }
-        gameHistory.dump(new Move(lastHint, lastMovement, treasurePos));
-        System.out.println("" +
-                "treasurePos: " + treasurePos +
-                "searcherPos:  " + searcherPos +
-                "");
-        stepSim++;
+    public void registerListener(Runnable runnable) {
+        views.add(runnable);
     }
 
     /**
-     * This simulates the whole game, until its finished.
+     * This will run the views, e.g. start the run() method.
      */
-    public void beat() {
-        while (!finished) {
-            move();
-        }
+    public void runListeners() {
+        views.forEach(runnable -> executorService.execute(runnable));
     }
 
     /**
-     * Simulates a fixed number of steps.
-     * Breaks, when the game is finished.
-     *
-     * @param steps number of steps
+     * @return The whole List of geometryItems of the gameHistory
      */
-    public void move(int steps) {
-        for (int i = 0; i < steps; i++) {
-            if (finished) {
-                break;
-            }
-            move();
-        }
-    }
-
-    /**
-     * @return whether the performed {@link Movement}' by the searcher and {@link Hint}'s from the hider were correct.
-     */
-    protected boolean checkConsistency() {
-        // TODO implement
-        // AngleHints must be correct
-        // AngleHints <=180 degrees
-        // CircleHints correct
-        // CircleHints must contain eachother !?
-        return true;
-    }
-
-    /**
-     * @return whether the searcher located the treasure successfully.
-     */
-    protected boolean located(List<GeometryItem<Point>> geometryItemsList) {
-        assert geometryItemsList.size() > 0;
-
-        // Did the searcher move ?
-        if (geometryItemsList.size() == 1) {
-            return geometryItemsList.get(0).getObject().distance(treasurePos) <= 1;
-        } else {
-            Point lastPoint = null;
-            for (GeometryItem<Point> geometryItem : geometryItemsList) {
-                Point point = geometryItem.getObject();
-                if (lastPoint == null) {
-                    lastPoint = point;
-                } else {
-                    // Check the gap of each move-segment and treasurePos
-                    LineSegment lineSegment = new LineSegment(new Coordinate(lastPoint.getX(), lastPoint.getY()),
-                            new Coordinate(point.getX(), point.getY()));
-                    // Usage of distancePerpendicular is completely incorrect here, since the line will be infinite
-                    if (lineSegment.distance(new Coordinate(treasurePos.getX(), treasurePos.getY())) <= 1) {
-                        return true;
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    /**
-     * register View-Threads.
-     * initialize searcher and hider.
-     * initialize searcher and treasure positions.
-     */
-    public void init() {
-        for (View view : view) {
-            gameHistory.registerListener(view);
-            view.init(gameHistory);
-        }
-
-        searcherPos = JTSUtils.createPoint(0, 0);
-        searcher.init(searcherPos);
-
-        treasurePos = hider.getTreasureLocation();
-        assert (treasurePos != null);
-
-        gameHistory.dump(new Move(
-                null,
-                new Movement(searcherPos),
-                treasurePos)
-        );
-
-        // Check, whether treasure spawns in range of searcher
-        List<GeometryItem<Point>> act = new ArrayList<>();
-        act.add(new GeometryItem<>(searcherPos, GeometryType.WAY_POINT));
-        if (located(act)) {
-            finish();
-        }
-    }
-
-    protected void finish() {
-        finished = true;
+    public List<GeometryItem> getGeometryItems() {
+        ArrayList<GeometryItem> geometryItems = moves.subList(0, stepView).stream()
+                .flatMap(move -> move.getGeometryItems().stream()).collect(Collectors.toCollection(ArrayList::new));
+        return geometryItems;
     }
 }
