+++ src/main/java/com/treasure/hunt/game/mods/insecurehints/InsecureGameEngine.java
@@ -1,0 +1,14 @@
+package com.treasure.hunt.game.mods.insecurehints;
+
+import com.treasure.hunt.game.GameEngine;
+import com.treasure.hunt.strategy.hider.Hider;
+import com.treasure.hunt.strategy.searcher.Searcher;
+import com.treasure.hunt.utils.Requires;
+
+@Requires(hider = InsecureHider.class, searcher = InsecureSearcher.class)
+public class InsecureGameEngine extends GameEngine {
+
+    public InsecureGameEngine(Searcher searcher, Hider hider) {
+        super(searcher, hider);
+    }
+}
