+++ src/main/java/com/treasure/hunt/game/mods/obstacles/ObstacleGameEngine.java
@@ -1,18 +1,15 @@
 package com.treasure.hunt.game.mods.obstacles;
 
-import com.treasure.hunt.game.GameManager;
+import com.treasure.hunt.game.GameEngine;
 import com.treasure.hunt.strategy.hider.Hider;
 import com.treasure.hunt.strategy.searcher.Searcher;
 import com.treasure.hunt.utils.Requires;
-import com.treasure.hunt.view.in_game.View;
-
-import java.util.List;
 
 @Requires(hider = ObstacleHider.class, searcher = ObstacleSearcher.class)
-public class ObstacleGameManager extends GameManager {
+public class ObstacleGameEngine extends GameEngine {
 
-    public ObstacleGameManager(Searcher searcher, Hider hider, List<View> view) {
-        super(searcher, hider, view);
+    public ObstacleGameEngine(Searcher searcher, Hider hider) {
+        super(searcher, hider);
     }
 
     @Override
