+++ src/main/java/com/treasure/hunt/view/in_game/View.java
@@ -1,12 +1,12 @@
 package com.treasure.hunt.view.in_game;
 
-import com.treasure.hunt.game.GameHistory;
+import com.treasure.hunt.game.GameManager;
 
 public interface View extends Runnable {
 
     /**
-     * @param gameHistory the {@link GameHistory}, the View gets its inputs.
+     * @param gameManager the {@link GameManager}, the View gets its inputs.
      */
-    void init(GameHistory gameHistory);
+    void init(GameManager gameManager);
 
 }
