+++ src/main/java/com/treasure/hunt/view/main/MainFrameController.java
@@ -1,12 +1,15 @@
 package com.treasure.hunt.view.main;
 
 
+import com.treasure.hunt.game.GameEngine;
 import com.treasure.hunt.game.GameManager;
 import com.treasure.hunt.strategy.hider.Hider;
 import com.treasure.hunt.strategy.searcher.Searcher;
+import com.treasure.hunt.view.in_game.View;
 import com.treasure.hunt.view.in_game.impl.CanvasView;
+import com.treasure.hunt.view.in_game.impl.DummyView;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
 public class MainFrameController {
@@ -27,13 +30,19 @@ public class MainFrameController {
         mainMenuController.show();
     }
 
-    public void onPlay(Class<? extends Searcher> searcherClass, Class<? extends Hider> hiderClass, Class<? extends GameManager> gameManager) throws Exception {
-        Searcher newSearcher = searcherClass.getDeclaredConstructor().newInstance();
-        Hider newHider = hiderClass.getDeclaredConstructor().newInstance();
+    public void onPlay(Class<? extends Searcher> searcherClass, Class<? extends Hider> hiderClass, Class<? extends GameEngine> gameEngineClass) throws Exception {
+        List<View> list = new ArrayList<>();
         CanvasView canvasView = new CanvasView();
-        GameManager gameManagerInstance = gameManager
-                .getDeclaredConstructor(Searcher.class, Hider.class, List.class)
-                .newInstance(newSearcher, newHider, Collections.singletonList(canvasView));
-        new CanvasController(canvasView, gameManagerInstance);
+        list.add(canvasView);
+        for (int i = 0; i < 7; i++) {
+            list.add(new DummyView());
+        }
+        /* TODO:
+         revert to this. We did this for testing purposes.
+         GameManager gameManager = new GameManager(searcherClass, hiderClass, gameEngineClass, Collections.singletonList(canvasView));
+        */
+        GameManager gameManager = new GameManager(searcherClass, hiderClass, gameEngineClass, list);
+
+        new CanvasController(canvasView, gameManager);
     }
 }
